refactor: Move folder structure
