+++ src/components/common/QuestionnaireLabel/LabelWrapper/QuestionnaireLabel.module.css
@@ -5,6 +5,7 @@
   top: 50%;
   right: 0;
   transform: translateY(-50%);
+  z-index: 9;
 }
 
 .label {
